Issue #SC-1486: Added a new samza job for Content cache updater
