+++ data-pipeline/content-cache-updater/src/main/config/content-cache-updater.properties
@@ -17,7 +17,7 @@
 
 # Job
 job.factory.class=org.apache.samza.job.yarn.YarnJobFactory
-job.name=__env__.UserCacheUpdater
+job.name=__env__.ContentCacheUpdater
 
 # YARN
 #yarn.package.path=file://${basedir}/target/${project.artifactId}-${pom.version}-distribution.tar.gz
@@ -26,8 +26,8 @@ yarn.package.path=http://__yarn_host__:__yarn_port__/__env__/${project.artifactI
 job.container.count=__redis_updater_yarn_container_count__
 
 # Task
-task.class=org.ekstep.ep.samza.task.UserCacheUpdaterTask
-task.inputs=kafka.__env__.telemetry.audit
+task.class=org.ekstep.ep.samza.task.ContentCacheUpdaterTask
+task.inputs=kafka.__env__.learning.graph.events
 task.checkpoint.factory=org.apache.samza.checkpoint.kafka.KafkaCheckpointManagerFactory
 task.checkpoint.system=kafka
 # Normally, this would be 3, but we have only one broker.
@@ -54,8 +54,6 @@ metrics.reporters=snapshot
 metrics.reporter.snapshot.class=org.apache.samza.metrics.reporter.MetricsSnapshotReporterFactory
 metrics.reporter.snapshot.stream=kafka.__env__.metrics
 output.metrics.topic.name=__env__.pipeline_metrics
-output.malformed.topic.name=__env__.telemetry.malformed
-input.audit.topic.name=__env__.telemetry.audit
 
 # Job Coordinator
 job.coordinator.system=kafka
@@ -78,20 +76,13 @@ redis.connection.idle.max=3
 redis.connection.idle.min=1
 redis.connection.minEvictableIdleTimeSeconds=120
 redis.connection.timeBetweenEvictionRunsSeconds=300
-redis.database.userStore.id=4
+redis.database.contentStore.id=5
+redis.database.dialCodeStore.id=6
 
 # content model fields to convert to List from String
 contentModel.fields.listType=gradeLevel,subject,medium,language
-user.signin.type.default = __user_sigin_type_default__
-user.login.type.default = __user_login_type_default__
-user.selfsignedin.typeList = google,self
-user.validated.typeList = sso
-user.self-siginin.key = Self-Signed-In
-user.valid.key = Validated
 
 # sunbird middleware cassandra config
 middleware.cassandra.host=__middleware_cassandra_host__
 middleware.cassandra.port=__middleware_cassandra_port__
 middleware.cassandra.keyspace=__middleware_cassandra_keyspace__
-middleware.cassandra.user_table=__middleware_cassandra_user_table__
-middleware.cassandra.location_table=__middleware_cassandra_location_table__
