+++ data-pipeline/content-cache-updater/src/main/java/org/ekstep/ep/samza/task/ContentCacheUpdaterSink.java
@@ -4,20 +4,13 @@ import org.apache.samza.task.MessageCollector;
 import org.ekstep.ep.samza.core.BaseSink;
 import org.ekstep.ep.samza.core.JobMetrics;
 
-public class UserCacheUpdaterSink extends BaseSink {
+public class ContentCacheUpdaterSink extends BaseSink {
 
-    public UserCacheUpdaterSink(MessageCollector collector, JobMetrics metrics) {
+    public ContentCacheUpdaterSink(MessageCollector collector, JobMetrics metrics) {
         super(collector, metrics);
     }
 
     public void success() {
         metrics.incSuccessCounter();
     }
-
-    public void error() {
-        metrics.incErrorCounter();
-    }
-
-    public void markSkipped() { metrics.incSkippedCounter(); }
-
 }
