+++ data-pipeline/user-cache-updater/src/test/java/org/ekstep/ep/samza/service/UserCacheUpdaterServiceTest.java
@@ -1,7 +1,6 @@
 package org.ekstep.ep.samza.service;
 
-import com.datastax.driver.core.Cluster;
-import com.datastax.driver.core.Row;
+import com.datastax.driver.core.exceptions.DriverException;
 import com.datastax.driver.core.querybuilder.QueryBuilder;
 import com.fiftyonred.mock_jedis.MockJedis;
 import com.google.gson.Gson;
@@ -27,10 +26,7 @@ import org.junit.Before;
 import org.junit.Test;
 import org.mockito.Mockito;
 import redis.clients.jedis.Jedis;
-import redis.clients.jedis.exceptions.JedisException;
 
-import javax.sql.rowset.RowSetFactory;
-import javax.sql.rowset.RowSetProvider;
 import java.lang.reflect.Type;
 import java.util.*;
 
@@ -55,10 +51,7 @@ public class UserCacheUpdaterServiceTest {
     private Counter counter;
     private Config configMock;
     private Integer userStoreId = 4;
-    private JedisException jedisException;
     private UserCacheUpdaterTask userCacheUpdaterTask;
-
-    private static final String MALFORMED_TOPIC = "telemetry.malformed";
 
     @Before
     public void setUp() {
@@ -66,15 +59,12 @@ public class UserCacheUpdaterServiceTest {
         userCacheUpdaterSinkMock = mock(UserCacheUpdaterSink.class);
         configMock = mock(Config.class);
         cassandraConnectMock = mock(CassandraConnect.class);
-//        cassandraConnectMock=new CassandraConnect("localhost",9042);
         contextMock = mock(TaskContext.class);
         jobMetricsMock = mock(JobMetrics.class);
         metricsRegistry = Mockito.mock(MetricsRegistry.class);
-        jedisException = mock(JedisException.class);
-        taskCoordinator=mock(TaskCoordinator.class);
-        messageCollector=mock(MessageCollector.class);
-        counter=mock(Counter.class);
-
+        taskCoordinator = mock(TaskCoordinator.class);
+        messageCollector = mock(MessageCollector.class);
+        counter = mock(Counter.class);
 
         stub(redisConnectMock.getConnection(userStoreId)).toReturn(jedisMock);
         envelopeMock = mock(IncomingMessageEnvelope.class);
@@ -96,29 +86,62 @@ public class UserCacheUpdaterServiceTest {
         stub(contextMock.getMetricsRegistry()).toReturn(metricsRegistry);
 
         userCacheUpdaterConfig=new UserCacheUpdaterConfig(configMock);
-//System.out.println(userCacheUpdaterConfig.userStoreDb());
         userCacheUpdaterService = new UserCacheUpdaterService(userCacheUpdaterConfig, redisConnectMock, cassandraConnectMock, jobMetricsMock);
         jedisMock.flushAll();
     }
 
     @Test
     public void shouldNotUpdateCacheForInvalidEvent() throws Exception {
+        stub(envelopeMock.getMessage()).toReturn(EventFixture.ANY_STRING);
         userCacheUpdaterTask = new UserCacheUpdaterTask(configMock, contextMock, cassandraConnectMock,redisConnectMock);
         userCacheUpdaterTask.process(envelopeMock,messageCollector,taskCoordinator);
-
-        stub(envelopeMock.getMessage()).toReturn(EventFixture.AUDIT_EVENT);
         UserCacheUpdaterSource source = new UserCacheUpdaterSource(envelopeMock);
         userCacheUpdaterService.process(source, userCacheUpdaterSinkMock);
-//        verify(userCacheUpdaterSinkMock, times(1)).error();
+
+        verify(userCacheUpdaterSinkMock, times(1)).error();
     }
 
     @Test
     public void shouldNotUpdateCacheForAnonymousUsers() throws Exception {
 
-        stub(envelopeMock.getMessage()).toReturn(EventFixture.AUDIT_EVENT2);
+        stub(envelopeMock.getMessage()).toReturn(EventFixture.AUDIT_EVENT3);
         UserCacheUpdaterSource source = new UserCacheUpdaterSource(envelopeMock);
         userCacheUpdaterService.process(source, userCacheUpdaterSinkMock);
-        verify(userCacheUpdaterSinkMock, times(1)).error();
+    }
+
+    @Test
+    public void shouldMarkEventSkippedForNullUserId() throws Exception {
+        stub(envelopeMock.getMessage()).toReturn(EventFixture.AUDIT_EVENT_NULL_USERID);
+        userCacheUpdaterTask = new UserCacheUpdaterTask(configMock, contextMock, cassandraConnectMock,redisConnectMock);
+        userCacheUpdaterTask.process(envelopeMock,messageCollector,taskCoordinator);
+        UserCacheUpdaterSource source = new UserCacheUpdaterSource(envelopeMock);
+        userCacheUpdaterService.process(source, userCacheUpdaterSinkMock);
+        verify(userCacheUpdaterSinkMock, times(1)).markSkipped();
+    }
+
+    @Test
+    public void testlocation() throws Exception {
+        stub(envelopeMock.getMessage()).toReturn(EventFixture.AUDIT_EVENT_LOGININTYPE);
+        UserCacheUpdaterSource source = new UserCacheUpdaterSource(envelopeMock);
+        jedisMock.set("3b46b4c9-3a10-439a-a2cb-feb5435b3a0d","{\"usersignintype\":\"google\"}");
+        String metadataQuery = QueryBuilder.select().all()
+                .from("sunbird", "user")
+                .where(QueryBuilder.eq("id", "3b46b4c9-3a10-439a-a2cb-feb5435b3a0d"))
+                .toString();
+        when(cassandraConnectMock.find(metadataQuery)).thenThrow(new DriverException("Cassandra Exception"));
+
+        userCacheUpdaterTask = new UserCacheUpdaterTask(configMock, contextMock, cassandraConnectMock,redisConnectMock);
+        userCacheUpdaterTask.process(envelopeMock,messageCollector,taskCoordinator);
+        userCacheUpdaterService.process(source, userCacheUpdaterSinkMock);
+    }
+
+    @Test public void shouldupdateLocationDetails() throws Exception {
+        stub(envelopeMock.getMessage()).toReturn(EventFixture.AUDIT_EVENT);
+        userCacheUpdaterTask = new UserCacheUpdaterTask(configMock, contextMock, cassandraConnectMock,redisConnectMock);
+        userCacheUpdaterTask.process(envelopeMock,messageCollector,taskCoordinator);
+
+        UserCacheUpdaterSource source = new UserCacheUpdaterSource(envelopeMock);
+        userCacheUpdaterService.process(source, userCacheUpdaterSinkMock);
     }
 
     @Test
@@ -127,11 +150,13 @@ public class UserCacheUpdaterServiceTest {
         stub(envelopeMock.getMessage()).toReturn(EventFixture.AUDIT_EVENT_SIGINTYPE);
         Gson gson = new Gson();
         String userId = "89490534-126f-4f0b-82ac-3ff3e49f3468";
-        jedisMock.set(userId, "{\"channel\":\"dikshacustodian\",\"phoneverified\":false}");
         UserCacheUpdaterSource source = new UserCacheUpdaterSource(envelopeMock);
         stub(envelopeMock.getSystemStreamPartition())
                 .toReturn(new SystemStreamPartition("kafka", "telemetry.audit", new Partition(1)));
         userCacheUpdaterService.process(source, userCacheUpdaterSinkMock);
+        Map<String, Object> userData = userCacheUpdaterService.updateUserCache(source.getEvent(), userId, userCacheUpdaterSinkMock);
+
+        jedisMock.set(userId, gson.toJson(userData));
 
         String cachedData = jedisMock.get(userId);
         Map<String, Object> parsedData = null;
@@ -149,11 +174,12 @@ public class UserCacheUpdaterServiceTest {
         stub(envelopeMock.getMessage()).toReturn(EventFixture.AUDIT_EVENT_SIGN_IN);
         Gson gson = new Gson();
         String userId = "89490534-126f-4f0b-82ac-3ff3e49f3468";
-        jedisMock.set(userId, "{\"channel\":\"dikshacustodian\",\"phoneverified\":false}");
         UserCacheUpdaterSource source = new UserCacheUpdaterSource(envelopeMock);
         stub(envelopeMock.getSystemStreamPartition())
                 .toReturn(new SystemStreamPartition("kafka", "telemetry.audit", new Partition(1)));
         userCacheUpdaterService.process(source, userCacheUpdaterSinkMock);
+        Map<String, Object> userData = userCacheUpdaterService.updateUserCache(source.getEvent(), userId, userCacheUpdaterSinkMock);
+        jedisMock.set(userId, gson.toJson(userData));
 
         String cachedData = jedisMock.get(userId);
         Map<String, Object> parsedData = null;
@@ -171,11 +197,12 @@ public class UserCacheUpdaterServiceTest {
         stub(envelopeMock.getMessage()).toReturn(EventFixture.AUDIT_EVENT_LOGININTYPE);
         Gson gson = new Gson();
         String userId = "3b46b4c9-3a10-439a-a2cb-feb5435b3a0d";
-        jedisMock.set(userId, "{\"channel\":\"dikshacustodian\",\"phoneverified\":false}");
         UserCacheUpdaterSource source = new UserCacheUpdaterSource(envelopeMock);
         stub(envelopeMock.getSystemStreamPartition())
                 .toReturn(new SystemStreamPartition("kafka", "telemetry.audit", new Partition(1)));
         userCacheUpdaterService.process(source, userCacheUpdaterSinkMock);
+        Map<String, Object> userData = userCacheUpdaterService.updateUserCache(source.getEvent(), userId, userCacheUpdaterSinkMock);
+        jedisMock.set(userId, gson.toJson(userData));
 
         String cachedData = jedisMock.get(userId);
         Map<String, Object> parsedData = null;
@@ -209,6 +236,17 @@ public class UserCacheUpdaterServiceTest {
     }
 
     @Test
+    public void shouldUpdateCache() throws Exception {
+        stub(envelopeMock.getMessage()).toReturn(EventFixture.AUDIT_EVENT_SIGINTYPE);
+        userCacheUpdaterTask = new UserCacheUpdaterTask(configMock, contextMock, cassandraConnectMock,redisConnectMock);
+        userCacheUpdaterTask.process(envelopeMock,messageCollector,taskCoordinator);
+
+        stub(envelopeMock.getMessage()).toReturn(EventFixture.AUDIT_EVENT_SIGIN_TYPE);
+        UserCacheUpdaterSource source = new UserCacheUpdaterSource(envelopeMock);
+        userCacheUpdaterService.process(source, userCacheUpdaterSinkMock);
+    }
+
+    @Test
     public void shouldUpdateCacheWithMetadataChangesAndLocationFORAUDIT() {
         stub(envelopeMock.getMessage()).toReturn(EventFixture.AUDIT_EVENT_METADATA_UPDATED);
         Gson gson = new Gson();
@@ -227,93 +265,5 @@ public class UserCacheUpdaterServiceTest {
         }
         assertEquals(parsedData.get("channel"), "dikshacustodian");
         verify(cassandraConnectMock, times(1)).find(anyString());
-    }
-
-    @Test
-    public void shouldUpdateCache() throws Exception {
-        stub(envelopeMock.getMessage()).toReturn(EventFixture.AUDIT_EVENT);
-
-
-//        row.ge
-
-//stub(row.get("loca"))
-//        stub(cassandraConnectMock.find("SELECT * FROM sunbird.user WHERE id='627a431d-4f5c-4adc-812d-1f01c5588555'"))
-//                .toReturn(Row);
-
-
-//        ArrayList<String> list = new ArrayList<>();
-//        list.add("6ffc5101-735d-43a3-a400-10f162361cd9");
-//        list.add("e70b2313-a73d-4f53-b0a1-7a5f6308a14d");
-//
-//
-//        ResultSet MockResultSet = mock(ResultSet.class);
-//        Row row = mock(Row.class);
-
-//        Map<String, Object> result = new HashMap<>();
-//        result.put("locationids",list);
-//        System.out.println(result);
-//
-//        List <Row> roeset = mock(List.class);
-//        roeset.add(0,row);
-
-//        Row mockRow = mock(Row.class, "locationStateid");
-//        Row mockDRow = mock(Row.class, "locationDistrictid");
-
-        Row mockRow = mock(Row.class, "6ffc5101-735d-43a3-a400-10f162361cd9");
-
-//        System.out.println(mockRow.getColumnDefinitions());
-        Row mockDRow = mock(Row.class, "e70b2313-a73d-4f53-b0a1-7a5f6308a14d");
-
-        List <Row> rowSet = new ArrayList<>();
-        rowSet.add(0, mockRow);
-        rowSet.add(1, mockDRow);
-
-        RowSetFactory aFactory = RowSetProvider.newFactory();
-        aFactory.createCachedRowSet();
-        System.out.println("cached row set factory: "+aFactory);
-
-
-//        System.out.println("row: "+rowSet);
-//        System.out.println(rowSet.get(0));
-
-//        String MetadataQuery = QueryBuilder.select().all()
-//                .from("sunbird", "user")
-//                .where(QueryBuilder.eq("id", "627a431d-4f5c-4adc-812d-1f01c5588555"))
-//                .toString();
-////        System.out.println(MetadataQuery);
-//
-//        System.out.println("cassandra mock connection: "+cassandraConnectMock);
-//        stub(cassandraConnectMock.find(MetadataQuery)).toReturn(rowSet);
-
-//        userCacheUpdaterTask = new UserCacheUpdaterTask(configMock, contextMock, cassandraConnectMock,redisConnectMock);
-//        userCacheUpdaterTask.process(envelopeMock,messageCollector,taskCoordinator);
-
-//        UserCacheUpdaterSource source = new UserCacheUpdaterSource(envelopeMock);
-//        userCacheUpdaterService.process(source, userCacheUpdaterSinkMock);
-//
-//
-//        verify(userCacheUpdaterSinkMock, times(1)).success();
-    }
-
-    @Test
-    public void shouldImproveCoverage() throws JedisException {
-        //    jedisMock.flushAll();
-        //      jedisMock.close();
-//        redisConnectMock.resetConnection();
-//        userCacheUpdaterService = new UserCacheUpdaterService(configMock, redisConnectMock, cassandraConnectMock, jobMetricsMock);
-
-//        stub(envelopeMock.getMessage()).toReturn(EventFixture.AUDIT_EVENT);
-//        String userId = "627a431d-4f5c-4adc-812d-1f01c5588555";
-//
-//        when(redisConnectMock.getConnection(userStoreId)).thenThrow(jedisException);
-
-//       when(jedisMock.get(userId)).thenThrow(JedisException.class);
-//        stub(jedisMock.get(userId)).toThrow(jedisException);
-//jedisMock.set(userId, "627a431d-4f5c-4adc-812d-1f01c5588555");
-
-        //  doThrow(new RuntimeException());
-
-//        UserCacheUpdaterSource source = new UserCacheUpdaterSource(envelopeMock);
-//        userCacheUpdaterService.process(source, userCacheUpdaterSinkMock);
     }
 }
